backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

sqlalchemy>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0
langsmith>=0.1.0

chromadb>=0.5.0

pypdf>=4.0.0
python-docx>=1.1.0

pydantic[email]>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.0
email-validator>=2.0.0
python-slugify>=8.0.0

structlog>=24.1.0

pytest>=8.0.0
pytest-asyncio>=0.23.0
